Modularize NumericCellFilter from CellFilteringPanel component
